ES
-Tarefas ligadas a projetos(funcionar)
-Tarefas individuais (funcionar)
-Copiar tarefa indivual para projeto (funcionar)
-Criar equiapas (funcionar)
-Convidar users(funcionar)
-Aceitar pedido (em desenvolvimento)

+++ ES2_webapp/DTO/TarefasIndCreate.cs
@@ -8,5 +8,5 @@ public class TarefasIndCreate
     public DateOnly? DtFim { get; set; }
     public TimeSpan HrFim { get; set; }
     public decimal PrecoHora { get; set; }
-    public decimal IdProjeto { get; set; }
+    public int IdProjeto { get; set; }
 }
